moved account table component
